backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "AI onboarding backend"
authors = [
    { name = "Craacky", email = "cbxj.ysa@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.133.0,<0.134.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "redis (>=7.2.0,<8.0.0)",
    "boto3 (>=1.42.56,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt>=4.0.1,<5.0.0",
    "pwdlib[argon2] (>=0.2.0, <1.0.0)",
    "email-validator (>=2.0.0, <2.3.0)",
    "python-multipart (>=0.0.9,<0.1.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
